requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argostranslate==1.7.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.0
botocore==1.25.9
branca==0.5.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
ckanapi==4.7
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cryptography==37.0.1
ctranslate2==2.17.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Fiona==1.8.18
flit_core==3.7.1
folium==0.12.1.post1
fonttools==4.33.3
frozenlist==1.3.0
GDAL==3.4.3
geojson==2.5.0
geopandas==0.10.2
ghp-import==2.1.0
google-trans-new==1.1.9
googletrans==4.0.0rc1
gql==3.2.0
gql-query-builder==0.1.7
graphql-core==3.2.1
h11==0.9.0
h2==3.2.0
handsdown==1.1.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
jinjasql==0.1.8
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
libretranslatepy==2.1.1
loguru==0.6.0
lxml==4.8.0
mapclassify==2.4.3
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.3.0
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
munkres==1.1.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathos==0.2.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.0.4
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycitylayers==0.0.1
pycparser==2.21
PyExecJS==1.5.1
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproj==3.3.1
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
Rtree==1.0.0
scikit-learn==1.1.0
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
setuptools==62.1.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
stanza==1.1.1
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
translate==3.6.1
translators==5.2.1
typed-ast==1.5.3
typing_extensions==4.2.0
unicodedata2==14.0.0
urllib3==1.26.9
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
wget==3.2
wheel==0.37.1
widgetsnbextension==3.6.0
xyzservices==2022.4.0
yarl==1.7.2
zipp==3.8.0
